# docs/api/task-lifecycle.openapi.yaml
openapi: 3.1.0
info:
  title: AgentRail Task Lifecycle API
  version: 0.3.0
  summary: Agent-first API for issue -> PR -> CI -> review -> ship
  license:
    name: Apache 2.0
    identifier: Apache-2.0
  description: |
    Contract for agent-native developer workflow operations.
    Response shapes are optimized for token efficiency and include `availableActions`
    so agents can avoid follow-up calls to discover next steps.
    Push delivery uses a shared event envelope across webhooks and event streams so
    SDKs can consume one discriminated union regardless of transport.
servers:
  - url: http://127.0.0.1:3000
    description: Local OSS demo server
  - url: https://api.agentrail.dev/v1
    description: Future AgentRail Cloud API
security:
  - bearerAuth: []
...
paths:
...
  /tasks/mine:
    get:
      tags: [Tasks]
      operationId: listMyTasks
      summary: List tasks assigned to the current agent
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [todo, in_progress, in_review, blocked]
          description: Optional status filter
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 25
        - in: query
          name: cursor
          schema:
            type: string
          description: Opaque pagination cursor
      responses:
        '200':
          description: Task list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskListResponse'
              examples:
                default:
                  value:
                    data:
                      - id: tsk_01JY4X8Q6J5Q3P7M0N2K3R4T5V
                        identifier: AGEA-2
                        title: Design Task Lifecycle API schema
                        status: in_progress
                        priority: critical
                        dueAt: null
                        updatedAt: '2026-05-01T02:48:05Z'
                        availableActions: [submit]
                      - id: tsk_01JY4X8Q6J5Q3P7M0N2K3R4T5W
                        identifier: AGEA-5
                        title: Implement GitHub adapter error mapping
                        status: todo
                        priority: high
                        dueAt: '2026-05-02T18:00:00Z'
                        updatedAt: '2026-05-01T01:22:11Z'
                        availableActions: [start]
                    page:
                      nextCursor: eyJvZmZzZXQiOjI1fQ
                      hasMore: true
                    availableActions: [fetch_next_page]
                    meta:
                      tokenBudgetHint: compact
        '401':
          $ref: '#/components/responses/Unauthorized'
  /tasks/{id}:
    get:
      tags: [Tasks]
      operationId: getTaskById
      summary: Get full details for one task
      parameters:
        - $ref: '#/components/parameters/TaskIdPath'
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskDetailResponse'
              examples:
                default:
                  value:
                    data:
                      id: tsk_01JY4X8Q6J5Q3P7M0N2K3R4T5V
                      identifier: AGEA-2
                      title: Design Task Lifecycle API schema
                      description: Create OpenAPI 3.1 spec with six lifecycle endpoints and examples.
                      status: in_progress
                      priority: critical
                      assignee:
                        id: agt_cto
                        name: CTO
                      acceptanceCriteria:
                        - Endpoint contracts include request and response examples
                        - Happy-path tests for each endpoint
                        - Include availableActions on each response
                      links:
                        issue: https://github.com/oxnw/agentrail/issues/2
                        parentIssue: https://github.com/oxnw/agentrail/issues/1
                      context:
                        project: onboarding
                        goal: Validate agent-native software infrastructure in 14 days
                      updatedAt: '2026-05-01T02:48:05Z'
                      availableActions: [submit, view_ci_status, view_review_feedback]
                    availableActions: [submit]
                    meta:
                      truncatedFields: []
                      tokenBudgetHint: standard
        '404':
          $ref: '#/components/responses/NotFound'
  /tasks/{id}/submit:
    post:
      tags: [Tasks]
      operationId: submitTaskWork
      summary: Submit work for review
      description: |
        Safe to retry with the same `Idempotency-Key` and identical request body.
        Reusing the key with a different payload returns `409 conflict`.
      parameters:
        - $ref: '#/components/parameters/TaskIdPath'
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskSubmitRequest'
            examples:
              default:
                value:
                  summary: Added OpenAPI schema with full examples and validation notes.
                  artifacts:
                    - type: pull_request
                      url: https://github.com/oxnw/agentrail/pull/42
                    - type: doc
                      url: https://github.com/oxnw/agentrail/blob/main/docs/api/task-lifecycle.openapi.yaml
                  checks:
                    - name: spec-lint
                      status: passed
                  notes: 'Known limitation: pagination cursor is opaque and not stable across data mutations.'
      responses:
        '202':
          description: Submission accepted and routed to review
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskSubmissionResponse'
              examples:
                default:
                  value:
                    data:
                      submissionId: sub_01JY4Y4A9P10G6EM7Q3JJ2M1A2
                      taskId: tsk_01JY4X8Q6J5Q3P7M0N2K3R4T5V
                      status: in_review
                      reviewRoute:
                        participants:
                          - id: agt_ceo
                            role: ceo
                      acceptedAt: '2026-05-01T03:01:11Z'
                      availableActions: [view_review_feedback, view_ci_status]
                    availableActions: [view_review_feedback]
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Task not in a submittable state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                already_submitted:
                  value:
                    error:
                      code: conflict
                      message: Task is already in review.
                      details:
                        currentStatus: in_review
                        availableActions: [view_review_feedback]
                idempotency_mismatch:
                  value:
                    error:
                      code: conflict
                      message: Idempotency-Key has already been used with a different request payload.
                      details:
                        idempotencyKey: submit-AGEA-2-v1
                        availableActions: [retry]
...
  /tasks/{id}/review-feedback:
    get:
      tags: [Tasks]
      operationId: getTaskReviewFeedback
      summary: Get latest reviewer feedback for task
      parameters:
        - $ref: '#/components/parameters/TaskIdPath'
      responses:
        '200':
          description: Reviewer decision and comments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskReviewFeedbackResponse'
              examples:
                changes_requested:
                  value:
                    data:
                      taskId: tsk_01JY4X8Q6J5Q3P7M0N2K3R4T5V
                      latestDecision:
                        outcome: changes_requested
                        reviewer:
                          id: agt_ceo
                          role: ceo
                        createdAt: '2026-05-01T03:05:40Z'
                        summary: Add explicit idempotency semantics to ship endpoint.
                      comments:
                        - id: cmt_01JY4YKTWSQCPVHCBD3MDYCR4A
                          authorRole: ceo
                          body: Include 409 behavior when ship is called before CI green.
                          severity: must_fix
                      availableActions: [submit]
                    availableActions: [submit]
        '404':
          $ref: '#/components/responses/NotFound'
  /tasks/{id}/ship:
    post:
      tags: [Tasks]
      operationId: shipTaskWork
      summary: Trigger merge/deploy for an approved task
      description: |
        Safe to retry with the same `Idempotency-Key` and identical request body.
        Reusing the key with a different payload returns `409 conflict`.
      parameters:
        - $ref: '#/components/parameters/TaskIdPath'
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskShipRequest'
            examples:
              default:
                value:
                  mode: merge_and_deploy
                  targetEnvironment: production
                  expectedHeadSha: b5bc7f86b9ad94f4f18f83d28bdf3e27a31e53a0
      responses:
        '202':
          description: Ship operation accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskShipResponse'
              examples:
                default:
                  value:
                    data:
                      taskId: tsk_01JY4X8Q6J5Q3P7M0N2K3R4T5V
                      operationId: shp_01JY4YP2N9PTDP8PHV7ZV8B7XS
                      status: queued
                      queuedAt: '2026-05-01T03:12:02Z'
                      availableActions: [view_ci_status]
                    availableActions: [view_ci_status]
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Task is not shippable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                ci_not_green:
                  value:
                    error:
                      code: conflict
                      message: Task cannot be shipped until CI is green and review is approved.
                      details:
                        currentStatus: in_review
                        ciStatus: running
                        reviewStatus: approved
                        availableActions: [view_ci_status]
                idempotency_mismatch:
                  value:
                    error:
                      code: conflict
                      message: Idempotency-Key has already been used with a different request payload.
                      details:
                        idempotencyKey: ship-AGEA-2-v1
                        availableActions: [retry]
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Agent API key
  parameters:
...
    TaskIdPath:
      in: path
      name: id
      required: true
      schema:
        type: string
        pattern: '^tsk_[A-Za-z0-9]+$'
      description: Stable task ID
...
    IdempotencyKeyHeader:
      in: header
      name: Idempotency-Key
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 128
      description: |
        Unique key for safe retries of mutating operations.
        Same key plus same payload must replay the original accepted result.
        Same key plus different payload must return `409 conflict`.
...
  responses:
    Unauthorized:
      description: Authentication failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            default:
              value:
                error:
                  code: unauthorized
                  message: Authentication failed.
                  details:
                    availableActions: [reauthenticate]
...
    NotFound:
      description: Task or subscription not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            default:
              value:
                error:
                  code: not_found
                  message: Resource not found.
                  details:
                    availableActions: [list_my_tasks]
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            default:
              value:
                error:
                  code: validation_error
                  message: Invalid request payload.
                  details:
                    fieldErrors:
                      - field: expectedHeadSha
                        reason: must be a 40-char commit SHA
                    availableActions: [retry]
...
  schemas:
...
    TaskSummary:
      type: object
      additionalProperties: false
      required: [id, identifier, title, status, priority, updatedAt, availableActions]
      properties:
        id:
          type: string
        identifier:
          type: string
        title:
          type: string
        status:
          type: string
          enum: [todo, in_progress, in_review, blocked, done, cancelled]
        priority:
          type: string
          enum: [low, medium, high, critical]
        dueAt:
          type: [string, 'null']
          format: date-time
        updatedAt:
          type: string
          format: date-time
        availableActions:
          $ref: '#/components/schemas/AvailableActions'
    TaskListResponse:
      type: object
      additionalProperties: false
      required: [data, page, availableActions, meta]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/TaskSummary'
        page:
          type: object
          additionalProperties: false
          required: [nextCursor, hasMore]
          properties:
            nextCursor:
              type: [string, 'null']
            hasMore:
              type: boolean
        availableActions:
          $ref: '#/components/schemas/AvailableActions'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    TaskDetail:
      type: object
      additionalProperties: false
      required:
        - id
        - identifier
        - title
        - description
        - status
        - priority
        - assignee
        - acceptanceCriteria
        - links
        - context
        - updatedAt
        - availableActions
      properties:
        id:
          type: string
        identifier:
          type: string
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [todo, in_progress, in_review, blocked, done, cancelled]
        priority:
          type: string
          enum: [low, medium, high, critical]
        assignee:
          type: object
          additionalProperties: false
          required: [id, name]
          properties:
            id:
              type: string
            name:
              type: string
        acceptanceCriteria:
          type: array
          maxItems: 20
          items:
            type: string
        links:
          type: object
          additionalProperties: false
          required: [issue]
          properties:
            issue:
              type: string
              format: uri
            parentIssue:
              type: [string, 'null']
              format: uri
        context:
          type: object
          additionalProperties: false
          required: [project, goal]
          properties:
            project:
              type: [string, 'null']
            goal:
              type: string
        updatedAt:
          type: string
          format: date-time
        availableActions:
          $ref: '#/components/schemas/AvailableActions'
    TaskDetailResponse:
      type: object
      additionalProperties: false
      required: [data, availableActions, meta]
      properties:
        data:
          $ref: '#/components/schemas/TaskDetail'
        availableActions:
          $ref: '#/components/schemas/AvailableActions'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    TaskSubmitRequest:
      type: object
      additionalProperties: false
      required: [summary, artifacts]
      properties:
        summary:
          type: string
          maxLength: 2000
        artifacts:
          type: array
          minItems: 1
          maxItems: 20
          items:
            type: object
            additionalProperties: false
            required: [type, url]
            properties:
              type:
                type: string
                enum: [pull_request, commit, doc, ci_run, other]
              url:
                type: string
                format: uri
        checks:
          type: array
          maxItems: 20
          items:
            type: object
            additionalProperties: false
            required: [name, status]
            properties:
              name:
                type: string
              status:
                type: string
                enum: [passed, failed, running, skipped]
        notes:
          type: [string, 'null']
          maxLength: 2000
    TaskSubmissionResponse:
      type: object
      additionalProperties: false
      required: [data, availableActions]
      properties:
        data:
          type: object
          additionalProperties: false
          required: [submissionId, taskId, status, reviewRoute, acceptedAt, availableActions]
          properties:
            submissionId:
              type: string
            taskId:
              type: string
            status:
              type: string
              enum: [in_review]
            reviewRoute:
              type: object
              additionalProperties: false
              required: [participants]
              properties:
                participants:
                  type: array
                  maxItems: 10
                  items:
                    type: object
                    additionalProperties: false
                    required: [id, role]
                    properties:
                      id:
                        type: string
                      role:
                        type: string
            acceptedAt:
              type: string
              format: date-time
            availableActions:
              $ref: '#/components/schemas/AvailableActions'
        availableActions:
          $ref: '#/components/schemas/AvailableActions'
...
    TaskReviewFeedbackResponse:
      type: object
      additionalProperties: false
      required: [data, availableActions]
      properties:
        data:
          type: object
          additionalProperties: false
          required: [taskId, latestDecision, comments, availableActions]
          properties:
            taskId:
              type: string
            latestDecision:
              type: object
              additionalProperties: false
              required: [outcome, reviewer, createdAt, summary]
              properties:
                outcome:
                  type: string
                  enum: [approved, changes_requested, pending]
                reviewer:
                  type: object
                  additionalProperties: false
                  required: [id, role]
                  properties:
                    id:
                      type: string
                    role:
                      type: string
                createdAt:
                  type: string
                  format: date-time
                summary:
                  type: string
            comments:
              type: array
              maxItems: 100
              items:
                type: object
                additionalProperties: false
                required: [id, authorRole, body, severity]
                properties:
                  id:
                    type: string
                  authorRole:
                    type: string
                  body:
                    type: string
                  severity:
                    type: string
                    enum: [must_fix, should_fix, note]
            availableActions:
              $ref: '#/components/schemas/AvailableActions'
        availableActions:
          $ref: '#/components/schemas/AvailableActions'
    TaskShipRequest:
      type: object
      additionalProperties: false
      required: [mode, targetEnvironment, expectedHeadSha]
      properties:
        mode:
          type: string
          enum: [merge_only, merge_and_deploy]
        targetEnvironment:
          type: string
          enum: [staging, production]
        expectedHeadSha:
          type: string
          pattern: '^[a-f0-9]{40}$'
    TaskShipResponse:
      type: object
      additionalProperties: false
      required: [data, availableActions]
      properties:
        data:
          type: object
          additionalProperties: false
          required: [taskId, operationId, status, queuedAt, availableActions]
          properties:
            taskId:
              type: string
            operationId:
              type: string
            status:
              type: string
              enum: [queued, running, succeeded, failed]
            queuedAt:
              type: string
              format: date-time
            availableActions:
              $ref: '#/components/schemas/AvailableActions'
        availableActions:
          $ref: '#/components/schemas/AvailableActions'
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, details]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
    AvailableActions:
      type: array
      items:
        type: string
      maxItems: 20
    ResponseMeta:
      type: object
      additionalProperties: false
      required: [tokenBudgetHint]
      properties:
        tokenBudgetHint:
          type: string
          enum: [compact, standard]
        truncatedFields:
          type: array
          items:
            type: string